Add files via upload
TEST#2 is code that models the magnus effect on projectile motion, using odeint to solve the set of partial differential equations, and matplotlib to plot the solutions.
Predator_Prey_3 is part of the final group project for the Computational Physics course. It models the Lokta-Volterra (Predator-Prey) interaction when food-pressures on the prey species is taken into account.
Stochastic_extra_credit is part of an in-class assignment. We were given code that models a simple 2-step nuclear decay using the stochastic method. The code here is an expansion of that code to include multiple different species, some with more than one mode of decay that needed to be accounted for. 
STBOO_ch is part of the astronomy research I worked on while at UT. The code imports data files containing data on the relative flux of light in the CH G-band from the star ST Bootes, and plots the measured spectrum (scatter) against synthetic spectra (
gray) that were developed using MOOG, which takes into account known stellar parameters and elemental abundances that influence the flux of light in the region being analyzed.